requirements.txt
alembic==1.5.3
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
curl==0.0.1
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gevent==21.1.2
greenlet==1.0.0
grequests==0.6.0
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.20.0
packaging==20.8
pandas==1.2.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.22
tinydb==4.3.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
